pyproject.toml
[tool.poetry]
name = "tcc"
version = "0.1.0"
description = ""
authors = ["GuilhermehVinicius <guilherme.ribeiro.business@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyterlab = "^4.3.4"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.37"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
